feat: Add new API for ISS data
The data from new API is a dictionary that I only need to add the rever-
se geolacated address to.

All conversions to human readable formats for e.g. the unix timestamp to
UTC will be done for frontend. Data to go in database as is

Remove `parse_iss_data`, `ISS_NOW_URL`, and writing to CSV for now. Next
steps include writing data to database.